requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
easyocr==1.7.1
requests==2.31.0
gunicorn==21.2.0